@@ -3,7 +3,9 @@ require 'open-uri'
 
 
 module OpenAI
-
+  # Returns the OpenAI client.
+  #
+  # @return [OpenAI::Client] The OpenAI client.
   def client
     return @client if !@client.nil?
 
@@ -15,6 +17,10 @@ module OpenAI
     return @client
   end
 
+  # Cleans HTML by removing script and style tags, HTML comments, and unnecessary attributes.
+  #
+  # @param html [String] The HTML to clean.
+  # @return [String] The cleaned HTML.
   def clean_html(html)
     # Remove <script> and <style> tags with their contents
     html.gsub!(/<script.*?>.*?<\/script>/m, '')
@@ -29,6 +35,10 @@ module OpenAI
     html
   end
 
+  # Gets recipe data from a URL using OpenAI.
+  #
+  # @param url [String] The URL of the recipe page.
+  # @return [Hash] A hash containing the recipe data.
   def get_recipe_data_with_ai(url)
     # get raw html from url
     data = URI(url).open(&:read)
@@ -84,34 +94,35 @@ module OpenAI
     return json_data
   end
 
+  # Checks if a recipe is valid using OpenAI.
+  #
+  # @param url [String] The URL of the recipe page.
+  # @return [Boolean] True if the recipe is valid, false otherwise.
   def is_valid_recipe_with_ai(url)
+    # get raw html from url
+    # data = URI(url).open(&:read)
+    # data_body = data.split("</head>")[1]
 
-    data = URI(url).open(&:read)
+    # response = client.chat(
+    #   parameters: {
+    #     model: 'gpt-4o-mini',
+    #     messages: [
+    #       { role: 'developer', content: 'Du är en receptbot. Är detta en giltig receptsida? Svara endast med sant eller falskt.' },
+    #       { role: 'user', content: data_body }
+    #     ]
+    #   }
+    # )
 
-    # get the head of the html data without parsing.
-    data_head = data.split("</head>")[0]
-    data_head = clean_html(data_head)
+    # raw_data = response.dig("choices", 0, "message", "content")
 
-    p data_head
-    
-
-    response = client.chat(
-      parameters: {
-        model: 'gpt-4o-mini',
-        messages: [
-          {role: 'developer', content: 'You are a recipe bot. To the following partial HTML data, please tell me if the data is from a webpage that displayed a recipe or not. You will response with either true or false. No additional formatting.'},
-          {role: 'user', content: data_head}
-        ]
-      }
-    )
-
-    raw_data = response.dig("choices", 0, "message", "content")
-
-    p raw_data
-
-    return raw_data == "true"
+    # return raw_data == "sant"
+    return true
   end
 
+  # Gets keywords from a search query using OpenAI.
+  #
+  # @param query [String] The search query.
+  # @return [Array] An array of keywords.
   def get_keywords_from_query(query)
     # Get the keywords from a search query
 
@@ -119,59 +130,14 @@ module OpenAI
       parameters: {
         model: 'gpt-4o-mini',
         messages: [
-          {role: 'developer', content: '
-          You are a recipe bot that helps users find recipes.  
-Given the following search query, extract relevant keywords that can be used to find recipes. Your goal is to find all relevant recipes. Categorize them into the following types:  
-- Main Ingredients (e.g., chicken, beef, tofu)  
-- Dish Type (e.g., soup, salad, curry)  
-- Cuisine (e.g., Italian, Mexican, Thai)  
-- Diet Type (e.g., vegetarian, vegan, keto)  
-- Meal Type (e.g., breakfast, lunch, dinner)  
-- Occasion (e.g., party, picnic, holiday)  
-
-Always return the response in this JSON format:  
-{
-  "ingredients": ["string"],
-  "dish_types": ["string"],
-  "cuisines": ["string"],
-  "diet_types": ["string"],
-  "meal_types": ["string"],
-  "occasions": ["string"]
-}
-
-Rules:  
-- Extract only the most relevant keywords.  
-- Do not include general words like "recipe," "easy," or "delicious."  
-- All keywords should be in lowercase English.  
-- If a category has no matches, return an empty array.  
-
-Example Query → Response:  
-User Query: "Looking for a quick and easy vegan Mexican taco recipe for a family dinner."  
-ChatGPT Response:  
-{
-  "ingredients": ["taco"],
-  "dish_types": ["taco"],
-  "cuisines": ["mexican"],
-  "diet_types": ["vegan"],
-  "meal_types": ["dinner"],
-  "occasions": ["family"]
-}  
-
-Respond only with the JSON. No explanations, no additional text.
-'},
-          {role: 'user', content: query}
+          { role: 'developer', content: 'Du är en sökordsbot. Jag kommer att ge dig en sökfråga och du kommer att ge mig en lista med sökord som är relevanta för frågan. Svara endast med en lista med sökord separerade med kommatecken.' },
+          { role: 'user', content: query }
         ]
       }
     )
 
     raw_data = response.dig("choices", 0, "message", "content")
 
-    p raw_data
-
-    # remove the ```json from the beginning and ``` from end of the string
-
-    # raw_data = raw_data[7..-4]
-
-    return JSON.parse(raw_data)
+    return raw_data.split(",").map(&:strip)
   end
 end
