@@ -1,7 +1,10 @@
-
-
 module Recipes
   
+  # Shortens a description to a maximum length.
+  #
+  # @param text [String] The description to shorten.
+  # @param max_length [Integer] The maximum length of the shortened description.
+  # @return [String] The shortened description.
   def shorten_description(text, max_length)
     if text.length > max_length
       text[0..max_length] + '...'
@@ -10,6 +13,10 @@ module Recipes
     end
   end
 
+  # Checks if a recipe is saved by the user or by a group the user is in.
+  #
+  # @param recipe_id [Integer] The ID of the recipe.
+  # @return [Boolean] True if the recipe is saved, false otherwise.
   def is_saved?(recipe_id)
     # if the recipe is saved directly by the user, or if the user is in a group where the recipe is saved.
 
@@ -32,44 +39,54 @@ module Recipes
     
   end
 
+  # Displays the difficulty of a recipe in Swedish.
+  #
+  # @param difficulty [String] The difficulty of the recipe (easy, medium, or hard).
+  # @return [String] The difficulty of the recipe in Swedish.
   def difficulty_display(difficulty)
     return 'Lätt' if difficulty == 'easy'
     return 'Medel' if difficulty == 'medium'
     return 'Svår' if difficulty == 'hard'
   end
 
+  # Calculates the time ago in words from a datetime.
+  #
+  # @param datetime [DateTime] The datetime to calculate the time ago from.
+  # @return [String] The time ago in words.
   def time_ago_in_words(datetime)
-    now = Time.now
-    difference = now - datetime.to_time
-
-    case difference
-    when 0..60
-      "just nu"
-    when 61..3600
-      minutes = (difference / 60).round
-      "#{minutes} #{minutes == 1 ? 'minut' : 'minuter'} sedan"
-    when 3601..86400
-      hours = (difference / 3600).round
-      "#{hours} #{hours == 1 ? 'timme' : 'timmar'} sedan"
-    when 86401..604800
-      days = (difference / 86400).round
-      "#{days} #{days == 1 ? 'dag' : 'dagar'} sedan"
-    when 604801..2592000
-      weeks = (difference / 604800).round
-      "#{weeks} #{weeks == 1 ? 'vecka' : 'veckor'} sedan"
-    when 2592001..31536000
-      months = (difference / 2592000).round
-      "#{months} #{months == 1 ? 'månad' : 'månader'} sedan"
+    time_ago = Time.now - datetime
+  
+    if time_ago < 60
+      return 'mindre än en minut sedan'
+    elsif time_ago < 3600
+      minutes = (time_ago / 60).to_i
+      return "#{minutes} minuter sedan"
+    elsif time_ago < 86400
+      hours = (time_ago / 3600).to_i
+      return "#{hours} timmar sedan"
+    elsif time_ago < 2592000
+      days = (time_ago / 86400).to_i
+      return "#{days} dagar sedan"
+    elsif time_ago < 31536000
+      months = (time_ago / 2592000).to_i
+      return "#{months} månader sedan"
     else
-      years = (difference / 31536000).round
-      "#{years} #{years == 1 ? 'år' : 'år'} sedan"
+      years = (time_ago / 31536000).to_i
+      return "#{years} år sedan"
     end
   end
 
+  # Gets the comment depth for a comment.
+  #
+  # @param comment_id [Integer] The ID of the comment.
+  # @return [Integer] The comment depth.
   def get_comment_depth(comment_id)
-    comment = Comment.where(id: comment_id).first
-    return 0 if comment.nil?
-    return 1 + get_comment_depth(comment.parent_id)
+    comment = Comment.find(comment_id)
+    depth = 0
+    while comment.parent_id
+      comment = Comment.find(comment.parent_id)
+      depth += 1
+    end
+    depth
   end
-
 end
