@@ -1,10 +1,12 @@
-
-
 class MyApp < Sinatra::Application
 
-  # simple search algorithm.
+  # @!group Routes
+
+  # Performs a search for recipes based on a query.
+  #
+  # @param q [String] The search query.
+  # @return [Haml] Rendered search results page.
   get '/search' do
-    
     @query = params[:q]
 
     if @query.nil? || @query.empty?
@@ -13,12 +15,7 @@ class MyApp < Sinatra::Application
 
     @recipes = get_recipes_from_query(@query)
 
-    
-    
-
-
-
     haml :'search/results'
   end
-
+  # @!endgroup
 end
